requirements.txt
urllib3==1.26.6
importlib-metadata==4.8.3
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
websockets_proxy==0.1.2

Pillow==8.0
django-statsd-mozilla==0.4.0
celery==4.4.7
django-celery-beat==2.0.0
web3==6.20.0
boto3==1.34
requests~=2.31.0
python-dateutil~=2.8.2
websockets==11.0.3
cachetools~=5.3.3
django-admin-csvexport==2.2
ortools~=9.10.4067
protobuf>=5.26.1
firebase-admin>=2.16.0
mixpanel>=4.10.1